挂代理

+++ src/main/java/com/example/hostinfo/util/FileUtil.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class FileUtil {
-    public static String resultsPath = CalabashUtil.path + "result";
+    public static String resultsPath =  CalabashUtil.path + "result";
     public static Logger logger = LoggerFactory.getLogger(FileUtil.class);
 
     public static BufferedReader reader;
@@ -26,15 +26,16 @@ public class FileUtil {
 
         if(file.isDirectory()){
             files = file.listFiles();
-
+            if(files.length == 0) {
+                return null;
+            }
             for (int i = 0; i < files.length; i++) {
                reader = new BufferedReader(new InputStreamReader(new FileInputStream(files[i].getPath())));
-                HostInfo hostInfo = readFile(reader);
+               HostInfo hostInfo = readFile(reader);
                if(null != hostInfo){
                    anchors.add(hostInfo);
                }
             }
-
             //读取完数据之后, 将文件删除掉
             for (int i = 0; i < files.length; i++) {
                 files[i].delete();
@@ -50,18 +51,22 @@ public class FileUtil {
      */
     private static HostInfo readFile(BufferedReader reader) throws IOException {
         HostInfo hostInfo = new HostInfo();
-        String []line1 = reader.readLine().split(",");
-        String []line2 = reader.readLine().split(",");
         try{
-            hostInfo.setCreateTime(new Timestamp(System.currentTimeMillis()));
-            hostInfo.setPlatform(Integer.parseInt(Task.getTypeByValue(line1[0].trim())));
-            hostInfo.setRoomId(line1[1].trim());
-            hostInfo.setFocusNum(Integer.parseInt(line2[0].split(":")[1]));
-            hostInfo.setHotNum(Integer.parseInt(line2[1].split(":")[1]));
-            hostInfo.setVipNum(Integer.parseInt(line2[2].split(":")[1]));
-        }catch (Exception e){
-            logger.error("转化失败 line1: " + line1 + "; line2: " + line2);
-            logger.error(e.getMessage());
+            String []line1 = reader.readLine().split(",");
+            String []line2 = reader.readLine().split(",");
+            try{
+                hostInfo.setCreateTime(new Timestamp(System.currentTimeMillis()));
+                hostInfo.setPlatform(Integer.parseInt(Task.getTypeByValue(line1[0].trim())));
+                hostInfo.setRoomId(line1[1].trim());
+                hostInfo.setFocusNum(Integer.parseInt(line2[0].split(":")[1]));
+                hostInfo.setHotNum(Integer.parseInt(line2[1].split(":")[1]));
+                hostInfo.setVipNum(Integer.parseInt(line2[2].split(":")[1]));
+            }catch (Exception e){
+                logger.error("转化失败 line1: " + line1 + "; line2: " + line2);
+                logger.error(e.getMessage());
+                return null;
+            }
+        }catch (Exception e) {
             return null;
         }
         reader.close();
